Cleaned up the renderForm routine, and made changes to show forms during a "print" intent
Changed the way annotations are hidden so that they still show during a
“print” intent, this way, choosing NOT to doForms means you can see
something where the forms should be still - useful for print or
previews. When setting doForm to true, you can still see html elements
because it uses the “display” intent.

Also reduced the size and complexity of renderForm, improved the
function to strip unneeded positional elements when controls fail to
create or other reasons to skip a field exist.