Removed dependency on .id and .name, fields contained inside form element and tweaker expanded to include page, canvas and form after creation
Replaced form container with a form element, rather than a div. Makes
more sense.

Removed dependency on .id and .name for fields. These can be added by
postCreationTweak if required, but adding them by default could cause
issues in the DOM and are not required for functionality.

.name remains in use on RADIO inputs, because this is how they are
grouped.

postCreationTweaker can be used to tweak form, canvas and page elements
as created (allows for adding no validate, auto-correct, post and so
forth).