Support for checkbox values (fixes issues with generating fdf)
Extract the correct options for a checkbox type and add to the data
stored with each field. This allows getting the correct true/false
values/constants rather than using a hard coded 1, booleans and the old
Yes/Off defaults (which is incorrect).

Expanded input param code to check for strings or booleans when it
comes to checkbox defaults.

Value attribute of checkboxes is now the correct value for the checked
state.